ai_agent_backend/requirements_agent.txt
fastapi
uvicorn
pydantic
langgraph
langchain
langchain-google-genai
supabase
python-dotenv
httpx
